+++ packages/cypress/src/add-commands/login.ts
@@ -1,15 +1,10 @@
 import semver from 'semver';
-import { v4 as uuidv4 } from 'uuid';
-import { buildOidcScope } from '@commercetools-frontend/application-shell/ssr';
 import {
   type ApplicationRuntimeEnvironment,
   CUSTOM_VIEW_HOST_ENTRY_POINT_URI_PATH,
+  GRAPHQL_TARGETS,
 } from '@commercetools-frontend/constants';
-import {
-  STORAGE_KEYS,
-  OIDC_RESPONSE_TYPES,
-  HTTP_STATUS_CODES,
-} from '../constants';
+import { HTTP_STATUS_CODES } from '../constants';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 declare const Cypress: any;
@@ -86,20 +81,39 @@ function isFeatureSupported(expectedVersion: string) {
   return semver.gte(Cypress.version, expectedVersion);
 }
 
+function isRunningOnLocalhost() {
+  const baseUrl = new URL(Cypress.config('baseUrl'));
+
+  return baseUrl.hostname === 'localhost';
+}
+
+const isCustomView = (commandOptions: CommandLoginOptions) =>
+  commandOptions.entryPointUriPath === CUSTOM_VIEW_HOST_ENTRY_POINT_URI_PATH;
+
 function loginByForm(commandOptions: CommandLoginOptions) {
-  if (isLocalhost()) {
-    throw new Error(
-      `At the moment, the "loginByForm" command only works when testing a Merchant Center production URL. Using form login in an application running on localhost is not supported due to issues with "cy.origin".`
-    );
-  }
+  const isLocalhost = isRunningOnLocalhost();
+  Cypress.log({
+    name: 'isLocalhost',
+    message: isLocalhost,
+  });
 
   const projectKey = commandOptions.projectKey ?? Cypress.env('PROJECT_KEY');
 
+  const isCustomViewConfigCommand = isCustomView(commandOptions);
+  const customEntityConfigCommand = isCustomViewConfigCommand
+    ? 'customViewConfig'
+    : 'customApplicationConfig';
+
+  const packageName =
+    (commandOptions as LoginToMerchantCenterForCustomViewCommandLoginOptions)
+      .packageName ?? Cypress.env('PACKAGE_NAME');
+
   cy.task(
-    'customApplicationConfig',
+    customEntityConfigCommand,
     {
       entryPointUriPath: commandOptions.entryPointUriPath,
       dotfiles: commandOptions.dotfiles,
+      ...(isCustomViewConfigCommand ? { packageName } : {}),
     },
     // Do not show log, as it may contain sensible information.
     { log: false }
@@ -112,7 +126,7 @@ function loginByForm(commandOptions: CommandLoginOptions) {
     // Log loaded application config for debugging purposes.
     Cypress.log({
       displayName: 'task',
-      name: 'customApplicationConfig',
+      name: customEntityConfigCommand,
       message: appConfig,
     });
 
@@ -125,26 +139,224 @@ function loginByForm(commandOptions: CommandLoginOptions) {
       userCredentials.email,
       commandOptions.entryPointUriPath,
     ];
-    // const mcUrl = new URL(appConfig.mcApiUrl);
-    // const mcFrontendHostname = mcUrl.hostname.replace('mc-api', 'mc');
 
+    /**
+     * The function is used by Cypress `session` command to store the
+     * browser state after executing the authentication flow.
+     */
     function authCallback() {
-      cy.visit(url, { onBeforeLoad: commandOptions.onBeforeLoad });
+      // Get the feature flags from the Merchant Center API so we can
+      // check whether Identity is enabled
+      cy.request({
+        method: 'POST',
+        url: `${appConfig.mcApiUrl}/graphql`,
+        body: {
+          operationName: 'AllFeatures',
+          query: `query AllFeatures { allFeatures { name value reason } }`,
+        },
+        headers: {
+          'content-type': 'application/json',
+          'x-graphql-target': GRAPHQL_TARGETS.MERCHANT_CENTER_BACKEND,
+          'x-graphql-operation-name': 'AllFeatures',
+          'x-project-key': projectKey,
+        },
+      }).then(
+        (res: {
+          body: {
+            data: {
+              allFeatures: { name: string; value: string; reason: string }[];
+            };
+          };
+        }) => {
+          const enableIdentity = res.body.data.allFeatures.find(
+            (feature) => feature.name === 'enableGlobalIdentity'
+          );
+          const isGlobalIdentityEnabled = Boolean(enableIdentity?.value);
+          Cypress.log({
+            name: 'isGlobalIdentityEnabled',
+            message: isGlobalIdentityEnabled,
+          });
+          const identityUrl =
+            Cypress.env('IDENTITY_URL') || 'https://identity.commercetools.com';
 
-      // NOTE: using `cy.origin` is currently disabled as it does not seem to properly work.
-      // Interestingly, starting an application locally using Vite works, however not when using Webpack.
-      // For now we keep it disabled until we find a solution.
+          // Visit the application URL, which triggers then the login flow.
+          cy.visit(url, {
+            onBeforeLoad: commandOptions.onBeforeLoad,
+          });
 
-      // // https://cypress.io/blog/2022/04/25/cypress-9-6-0-easily-test-multi-domain-workflows-with-cy-origin/
-      // cy.origin(
-      //   mcFrontendHostname,
-      //   { args: userCredentials },
-      //   fillLoginForm
-      // );
-      fillLoginForm(userCredentials);
+          /**
+           * There are different scenarios and variations on the flow depending
+           * on the environment (localhost, production) and if Identity is enabled.
+           *
+           * # When the application runs on localhost:3001.
+           *
+           * ## When Identity is enabled
+           * - The test visits the application at localhost:3001.
+           * - Cypress registers the default origin being localhost:3001.
+           * - The application is redirected to the Identity URL.
+           * - Cypress interacts with the Identity URL via `cy.origin`.
+           * - At the end of the flow, the test interacts with the application at localhost:3001.
 
-      // Wait for the route to be loaded so that the session can be saved.
-      cy.url().should('include', url);
+           * ## When Identity is disabled
+           * - The test visits the application at localhost:3001.
+           * - Cypress registers the default origin being localhost:3001.
+           * - The application is redirected to the Merchant Center login page.
+           * - Cypress interacts with the Merchant Center login page via `cy.origin`.
+           * - At the end of the flow, the test interacts with the application at localhost:3001.
+           *
+           * # When the application runs on a production URL.
+           *
+           * ## When Identity is enabled
+           * - The test visits the application at the production URL.
+           * - The MC Proxy performs a server-side redirect to the Identity URL.
+           * - Cypress registers the default origin being the Identity URL.
+           * - Cypress interacts with the Identity URL.
+           * - At the end of the flow, the test interacts with the application at the production URL via `cy.origin`.
+           *
+           * ## When Identity is disabled
+           * - The test visits the application at the production URL.
+           * - The MC Proxy renders the Merchant Center login page.
+           * - Cypress registers the default origin being the Merchant Center URL.
+           * - Cypress interacts with the Merchant Center login page.
+           * - At the end of the flow, the test interacts with the application at the production URL.
+           */
+
+          if (isLocalhost) {
+            if (isGlobalIdentityEnabled) {
+              // Cypress gets confused when switching between origins, likely because
+              // the redirect to Identity doesn't happen immediately.
+              // If we don't wait, Cypress fails as it tries to interact with `cy.origin`
+              // but the test is still in the initial origin URL.
+              // This is a bit unexpected and to be considered a workaround.
+              // eslint-disable-next-line cypress/no-unnecessary-waiting
+              cy.wait(1000);
+
+              // Use cy.origin to handle the identity domain
+              cy.origin(
+                identityUrl,
+                {
+                  args: {
+                    userCredentials,
+                    identityUrl,
+                  },
+                },
+                ({
+                  userCredentials,
+                  identityUrl,
+                }: {
+                  userCredentials: LoginCredentials;
+                  identityUrl: string;
+                }) => {
+                  cy.url().should('include', `${identityUrl}/login`);
+                  // Fill in the email and click Next
+                  cy.get('input[name="identifier"]').type(
+                    userCredentials.email
+                  );
+                  cy.get('button').contains('Next').click();
+
+                  // Wait for the password form to appear
+                  cy.get('input[name="password"]').should('be.visible');
+                  // Fill in the password and submit
+                  cy.get('input[name="password"]').type(
+                    userCredentials.password,
+                    { log: false }
+                  );
+                  cy.get('button').contains('Submit').click();
+                }
+              );
+
+              // Wait for the flow to redirect back to the application.
+              cy.get('[role="main"]').should('exist');
+              cy.url().should('include', url);
+            } else {
+              const mcUrl = appConfig.mcApiUrl.replace('mc-api', 'mc');
+              // See similar comment above regarding the usage of `cy.wait`.
+              // eslint-disable-next-line cypress/no-unnecessary-waiting
+              cy.wait(1000);
+              cy.origin(
+                mcUrl,
+                {
+                  args: {
+                    userCredentials,
+                    mcUrl,
+                  },
+                },
+                ({
+                  userCredentials,
+                  mcUrl,
+                }: {
+                  userCredentials: LoginCredentials;
+                  mcUrl: string;
+                }) => {
+                  cy.url().should('include', `${mcUrl}/login`);
+
+                  // Same as `fillLegacyLoginFormWithRetry`.
+                  // eslint-disable-next-line cypress/unsafe-to-chain-command
+                  cy.get('input[name=email]')
+                    // We use `force` as the MC login UI (production) in tests renders the
+                    // cookie banner overlapping the input fields.
+                    // To allow Cypress to interact with the input fields, we use `force`.
+                    .clear({ force: true })
+                    .type(userCredentials.email, { force: true });
+                  // eslint-disable-next-line cypress/unsafe-to-chain-command
+                  cy.get('input[name=password]')
+                    // We use `force` as the MC login UI (production) in tests renders the
+                    // cookie banner overlapping the input fields.
+                    // To allow Cypress to interact with the input fields, we use `force`.
+                    .clear({ force: true })
+                    .type(userCredentials.password, {
+                      log: false,
+                      force: true,
+                    });
+                  cy.get('button').contains('Sign in').click({ force: true });
+                }
+              );
+              // eslint-disable-next-line cypress/no-unnecessary-waiting
+              cy.wait(1000);
+
+              // Wait for the flow to redirect back to the application.
+              cy.get('[role="main"]').should('exist');
+              cy.url().should('include', url);
+            }
+          } else {
+            if (isGlobalIdentityEnabled) {
+              cy.url().should('include', `${identityUrl}/login`);
+              // Fill in the email and click Next
+              cy.get('input[name="identifier"]').type(userCredentials.email);
+              cy.get('button').contains('Next').click();
+
+              // Wait for the password form to appear
+              cy.get('input[name="password"]').should('be.visible');
+              // Fill in the password and submit
+              cy.get('input[name="password"]').type(userCredentials.password, {
+                log: false,
+              });
+              cy.get('button').contains('Submit').click();
+
+              // Wait for the flow to redirect back to the application.
+              cy.origin(
+                Cypress.config('baseUrl'),
+                {
+                  args: {
+                    url,
+                  },
+                },
+                ({ url }: { url: string }) => {
+                  cy.get('[role="main"]').should('exist');
+                  cy.url().should('include', url);
+                }
+              );
+            } else {
+              // Legacy login flow.
+              fillLegacyLoginFormWithRetry(userCredentials);
+
+              // Wait for the flow to redirect back to the application.
+              cy.get('[role="main"]').should('exist');
+              cy.url().should('include', url);
+            }
+          }
+        }
+      );
     }
 
     // For backwards compatibility.
@@ -180,155 +392,10 @@ function loginByForm(commandOptions: CommandLoginOptions) {
   });
 }
 
-const isCustomView = (commandOptions: CommandLoginOptions) =>
-  commandOptions.entryPointUriPath === CUSTOM_VIEW_HOST_ENTRY_POINT_URI_PATH;
-
-function loginByOidc(
-  commandOptions: CommandLoginOptions &
-    LoginToMerchantCenterForCustomViewCommandLoginOptions
-) {
-  const isCustomViewConfigCommand = isCustomView(commandOptions);
-  if (!isLocalhost()) {
-    throw new Error(
-      `The "loginByOidc" command only works when testing a Custom ${
-        isCustomViewConfigCommand ? 'View' : 'Application'
-      } running on localhost.`
-    );
-  }
-
-  const sessionNonce = uuidv4();
-  let projectKey: string | undefined = undefined;
-  if (commandOptions.entryPointUriPath !== 'account') {
-    projectKey = commandOptions.projectKey ?? Cypress.env('PROJECT_KEY');
-  }
-
-  const customEntityConfigCommand = isCustomViewConfigCommand
-    ? 'customViewConfig'
-    : 'customApplicationConfig';
-
-  const packageName = commandOptions.packageName ?? Cypress.env('PACKAGE_NAME');
-
-  if (isCustomViewConfigCommand && !packageName) {
-    throw new Error(
-      `Missing required option "packageName" when using the "loginToMerchantCenterForCustomView" command.`
-    );
-  }
-
-  cy.task(
-    customEntityConfigCommand,
-    {
-      entryPointUriPath: commandOptions.entryPointUriPath,
-      dotfiles: commandOptions.dotfiles,
-      ...(isCustomViewConfigCommand ? { packageName } : {}),
-    },
-    // Do not show log, as it may contain sensible information.
-    { log: false }
-  ).then((appConfig: ApplicationRuntimeEnvironment) => {
-    // Log loaded application config for debugging purposes.
-    Cypress.log({
-      displayName: 'task',
-      name: customEntityConfigCommand,
-      message: appConfig,
-    });
-
-    const applicationId = appConfig.applicationId;
-    const sessionScope = buildOidcScope({
-      projectKey,
-      oAuthScopes: appConfig.__DEVELOPMENT__?.oidc?.oAuthScopes,
-      additionalOAuthScopes:
-        appConfig.__DEVELOPMENT__?.oidc?.additionalOAuthScopes,
-      teamId: appConfig.__DEVELOPMENT__?.oidc?.teamId,
-      applicationId: appConfig.__DEVELOPMENT__?.oidc?.applicationId,
-    });
-    const userCredentials = commandOptions.login ?? {
-      email: Cypress.env('LOGIN_EMAIL') || Cypress.env('LOGIN_USER'),
-      password: Cypress.env('LOGIN_PASSWORD'),
-    };
-    // Perform the login using the API, then store some required values into the browser storage
-    // and redirect to the auth callback route.
-    const requestOptions = {
-      method: 'POST',
-      url: `${appConfig.mcApiUrl}/tokens`,
-      body: {
-        ...userCredentials,
-        client_id: applicationId,
-        response_type: OIDC_RESPONSE_TYPES.ID_TOKEN,
-        scope: sessionScope,
-        state: sessionNonce,
-        nonce: sessionNonce,
-      },
-      followRedirect: false,
-    };
-    cy.request(requestOptions).then((res: { body: { redirectTo: string } }) => {
-      const sessionKey = [
-        'loginByOidc',
-        userCredentials.email,
-        commandOptions.entryPointUriPath,
-      ];
-
-      function authCallback() {
-        cy.visit(res.body.redirectTo, {
-          onBeforeLoad(win: Window) {
-            if (projectKey) {
-              win.localStorage.setItem(
-                STORAGE_KEYS.ACTIVE_PROJECT_KEY,
-                projectKey
-              );
-            }
-            win.sessionStorage.setItem(
-              `${STORAGE_KEYS.NONCE}_${sessionNonce}`,
-              JSON.stringify({ applicationId, query: {} })
-            );
-            win.sessionStorage.setItem(
-              STORAGE_KEYS.SESSION_SCOPE,
-              sessionScope
-            );
-            if (commandOptions.onBeforeLoad) {
-              commandOptions.onBeforeLoad(win);
-            }
-          },
-        });
-        // Wait for the application to be loaded so that the session can be saved.
-        cy.get('#app-loader').should('not.exist');
-      }
-
-      // For backwards compatibility.
-      if (
-        isFeatureSupported('12.0.0') ||
-        Cypress.config('experimentalSessionAndOrigin')
-      ) {
-        // https://www.cypress.io/blog/2021/08/04/authenticate-faster-in-tests-cy-session-command/
-        cy.session(
-          sessionKey,
-          authCallback,
-          isFeatureSupported('10.9.0')
-            ? {
-                cacheAcrossSpecs:
-                  typeof commandOptions.disableCacheAcrossSpecs === 'boolean'
-                    ? !commandOptions.disableCacheAcrossSpecs
-                    : true,
-              }
-            : undefined
-        );
-      } else {
-        cy.log(
-          `We recommend to use "cy.session" to reduce the time to log in between tests. Make sure to have at least Cypress v12 or enable it via "experimentalSessionAndOrigin" for older Cypress versions.`
-        );
-        authCallback();
-      }
-
-      if (commandOptions.initialRoute) {
-        cy.visit(`${Cypress.config('baseUrl')}${commandOptions.initialRoute}`);
-        cy.url().should('include', commandOptions.initialRoute);
-      }
-    });
-  });
-}
-
 /* Utilities */
 
-const maxLoginAttempts = Cypress.config('maxLoginAttempts') ?? 3;
-function fillLoginForm(userCredentials: LoginCredentials) {
+const legacyMaxLoginAttempts = Cypress.config('maxLoginAttempts') ?? 3;
+function fillLegacyLoginFormWithRetry(userCredentials: LoginCredentials) {
   // Intercept the login request so we can retry it if we receive a TOO_MANY_REQUESTS status code
   cy.intercept('POST', '**/tokens').as('loginRequest');
 
@@ -349,12 +416,17 @@ function fillLoginForm(userCredentials: LoginCredentials) {
     cy.log(`Attempts left: ${attemptsLeft}`);
 
     // eslint-disable-next-line cypress/unsafe-to-chain-command
-    cy.get('input[name=email]').clear().type(userCredentials.email);
+    cy.get('input[name=email]')
+      .clear({ force: true })
+      .type(userCredentials.email, { force: true });
     // eslint-disable-next-line cypress/unsafe-to-chain-command
-    cy.get('input[name=password]').clear().type(userCredentials.password, {
-      log: false,
-    });
-    cy.get('button').contains('Sign in').click();
+    cy.get('input[name=password]')
+      .clear({ force: true })
+      .type(userCredentials.password, {
+        log: false,
+        force: true,
+      });
+    cy.get('button').contains('Sign in').click({ force: true });
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     cy.wait('@loginRequest').then((interception: any) => {
@@ -371,13 +443,7 @@ function fillLoginForm(userCredentials: LoginCredentials) {
     });
   }
 
-  attemptLogin(maxLoginAttempts);
+  attemptLogin(legacyMaxLoginAttempts);
 }
 
-function isLocalhost() {
-  const baseUrl = new URL(Cypress.config('baseUrl'));
-
-  return baseUrl.hostname === 'localhost';
-}
-
-export { loginByForm, loginByOidc, isLocalhost };
+export { loginByForm };