+++ packages/create-mc-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @commercetools-frontend/create-mc-app
 
+## 24.4.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-frontend/application-config@24.4.0
+
+## 24.3.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-frontend/application-config@24.3.0
+
 ## 24.2.1
 
 ### Patch Changes