+++ packages/l10n/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @commercetools-frontend/l10n
 
+## 24.4.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-frontend/sentry@24.4.0
+
+## 24.3.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-frontend/sentry@24.3.0
+
 ## 24.2.1
 
 ### Patch Changes