+++ packages/mc-dev-authentication/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @commercetools-frontend/mc-dev-authentication
 
+## 24.4.0
+
+## 24.3.0
+
 ## 24.2.1
 
 ## 24.2.0